[CoreBundle] Add UserRepository [UserBundle] Move function which has connection between user and customer to core UserRepository
