+++ src/Sylius/Bundle/UserBundle/Doctrine/ORM/UserRepository.php
@@ -23,63 +23,6 @@ class UserRepository extends EntityRepository implements UserRepositoryInterface
     /**
      * {@inheritdoc}
      */
-    public function createFilterPaginator(array $criteria = null, array $sorting = null)
-    {
-        $queryBuilder = $this->createQueryBuilder('o');
-
-        if (isset($criteria['query'])) {
-            $queryBuilder
-                ->leftJoin('o.customer', 'customer')
-                ->where('customer.emailCanonical LIKE :query')
-                ->orWhere('customer.firstName LIKE :query')
-                ->orWhere('customer.lastName LIKE :query')
-                ->orWhere('o.username LIKE :query')
-                ->setParameter('query', '%'.$criteria['query'].'%')
-            ;
-        }
-        if (isset($criteria['enabled'])) {
-            $queryBuilder
-                ->andWhere('o.enabled = :enabled')
-                ->setParameter('enabled', $criteria['enabled'])
-            ;
-        }
-
-        if (empty($sorting)) {
-            if (!is_array($sorting)) {
-                $sorting = [];
-            }
-            $sorting['updatedAt'] = 'desc';
-        }
-
-        $this->applySorting($queryBuilder, $sorting);
-
-        return $this->getPaginator($queryBuilder);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findForDetailsPage($id)
-    {
-        $queryBuilder = $this->createQueryBuilder('o');
-        $queryBuilder
-            ->leftJoin('o'.'.customer', 'customer')
-            ->addSelect('customer')
-            ->where($queryBuilder->expr()->eq('o'.'.id', ':id'))
-            ->setParameter('id', $id)
-        ;
-
-        $result = $queryBuilder
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-
-        return $result;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function countBetweenDates(\DateTime $from, \DateTime $to, $status = null)
     {
         $queryBuilder = $this->createQueryBuilder('o');
@@ -145,8 +88,7 @@ class UserRepository extends EntityRepository implements UserRepositoryInterface
         $queryBuilder = $this->createQueryBuilder('o');
 
         $queryBuilder
-            ->leftJoin('o'.'.customer', 'customer')
-            ->andWhere($queryBuilder->expr()->eq('customer.emailCanonical', ':email'))
+            ->andWhere($queryBuilder->expr()->eq('o.emailCanonical', ':email'))
             ->setParameter('email', $email)
         ;
